Moved yield
